Split parser code out of RegEx and LogicEx.
