+++ src/main/java/edu/washington/cs/knowitall/logic/LogicExpression.java
@@ -6,20 +6,17 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Scanner;
 import java.util.Stack;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import com.google.common.base.Predicate;
 import com.google.common.base.Function;
-import com.google.common.collect.Lists;
+import com.google.common.base.Predicate;
 
+import edu.washington.cs.knowitall.logic.Expression.Apply;
+import edu.washington.cs.knowitall.logic.Expression.Arg;
+import edu.washington.cs.knowitall.logic.Expression.Op;
+import edu.washington.cs.knowitall.logic.Expression.Paren;
 import edu.washington.cs.knowitall.logic.LogicException.ApplyLogicException;
 import edu.washington.cs.knowitall.logic.LogicException.CompileLogicException;
 import edu.washington.cs.knowitall.logic.LogicException.TokenizeLogicException;
-import edu.washington.cs.knowitall.logic.Expression.Apply;
-import edu.washington.cs.knowitall.logic.Expression.Op;
-import edu.washington.cs.knowitall.logic.Expression.Arg;
-import edu.washington.cs.knowitall.logic.Expression.Paren;
 
 /**
  * A logic expression engine that operates over user specified objects.
@@ -28,7 +25,7 @@ import edu.washington.cs.knowitall.logic.Expression.Paren;
  *
  * @param  <E>  the type of the base expressions
  */
-abstract public class LogicExpression<E> implements Predicate<E> {
+public class LogicExpression<E> implements Predicate<E> {
     private final Apply<E> expression;
 
     /***
@@ -37,93 +34,23 @@ abstract public class LogicExpression<E> implements Predicate<E> {
      * @throws TokenizeLogicException
      * @throws CompileLogicException
      */
-    public LogicExpression(String input)
+    protected LogicExpression(List<Expression<E>> expressions)
             throws TokenizeLogicException, CompileLogicException {
-        // convert to tokens
-        List<Expression<E>> tokens = tokenize(input);
-
         // put in reverse polish notation
-        List<Expression<E>> rpn = rpn(tokens);
+        List<Expression<E>> rpn = rpn(expressions);
 
         // compile the expression
-        expression = compile(rpn);
+        expression = buildAst(rpn);
     }
 
     /***
-     * The factory method creates an argument from the supplied token string.
-     * @param  argument  a string representation of a token
-     * @return  an evaluatable representation of a token
+     * Compile an infix list of tokens into an expression tree.
+     * @param rpn a list of tokens in infix form.
+     * @return an expression tree.
      */
-    public abstract Arg<E> factory(String argument);
-
-    public final static Pattern doubleQuoteStringLiteralRegex =
-            Pattern.compile("\"" + "((?:[^\"\\p{Cntrl}\\\\]|\\\\[\\\\'\"bfnrt]|\\\\u[a-fA-F0-9]{4})*)" + "\"");
-    public final static Pattern singleQuoteStringLiteralRegex =
-            Pattern.compile("'" + "([^']*)" + "'");
-    public final static Pattern regexLiteralRegex =
-            Pattern.compile("/" + "((?:[^/\\\\]*(?:\\\\)*(?:\\\\/)*)*)" + "/");
-    private final static List<Pattern> literalPatterns = Lists.newArrayList(
-            doubleQuoteStringLiteralRegex, singleQuoteStringLiteralRegex,
-            regexLiteralRegex);
-
-    /***
-     * The readToken method reads a token from the remaining LogicExpression string.
-     *
-     * A token may contain a string.  If it contains parentheses, the token
-     * will last until the parentheses are balanced.  And &, |, or unbalanced )
-     * will mark the end of a token.
-     *
-     * This is a default implementation that may be overriden.
-     * @param  remainder  the remaining text to tokenize
-     * @return  a token from the beginning on `remaining`
-     */
-    public String readToken(String remainder) {
-        final String token;
-        try {
-            Stack<Character> parens = new Stack<Character>();
-
-            int nextExpression;
-            for (nextExpression = 0; nextExpression < remainder.length(); nextExpression++) {
-                char c = remainder.charAt(nextExpression);
-
-                // check for quotation
-                String match = null;
-                for (Pattern pattern : literalPatterns) {
-                    Matcher matcher = pattern.matcher(remainder).region(
-                            nextExpression, remainder.length());
-                    if (matcher.lookingAt()) {
-                        match = matcher.group(0);
-                        break;
-                    }
-                }
-
-                if (match != null) {
-                    // we found and can consume a quotation
-                    nextExpression += match.length() - 1;
-                } else if (c == '(') {
-                    parens.push(c);
-                } else if (c == ')') {
-                    if (parens.isEmpty()) {
-                        break;
-                    } else {
-                        parens.pop();
-                    }
-                } else if (c == '&' || c == '|') {
-                    break;
-                }
-            }
-
-            token = remainder.substring(0, nextExpression).trim();
-        } catch (Exception e) {
-            throw new TokenizeLogicException("Error parsing token: "
-                    + remainder, e);
-        }
-
-        if (token.isEmpty()) {
-            throw new TokenizeLogicException("zero-length token found.");
-        }
-
-        return token;
+    public static <E> LogicExpression<E> compile(
+            final List<Expression<E>> expressions) {
+        return new LogicExpression<E>(expressions);
     }
 
     /***
@@ -134,12 +61,12 @@ abstract public class LogicExpression<E> implements Predicate<E> {
      */
     public static <E> LogicExpression<E> compile(final String input,
             final Function<String, Arg<E>> factoryDelegate) {
-        return new LogicExpression<E>(input) {
+        return new LogicExpressionParser<E>() {
             @Override
             public Arg<E> factory(String argument) {
                 return factoryDelegate.apply(argument);
             }
-        };
+        }.parse(input);
     }
 
     @Override
@@ -176,7 +103,7 @@ abstract public class LogicExpression<E> implements Predicate<E> {
      * @param rpn a list of tokens in infix form.
      * @return an expression tree.
      */
-    public static <E> Apply<E> compile(List<Expression<E>> rpn) {
+    public static <E> Apply<E> buildAst(List<Expression<E>> rpn) {
         if (rpn.isEmpty()) {
             return null;
         }
@@ -263,54 +190,6 @@ abstract public class LogicExpression<E> implements Predicate<E> {
     }
 
     /***
-     * Convert an infix string logic representation to an infix list of tokens.
-     * @param input an infix string logic representation.
-     * @param factory a delegate that converts a string representation of an
-     * argument into a token object.  @return
-     *
-     * @throws TokenizeLogicException
-     */
-    public List<Expression<E>> tokenize(String input)
-    throws TokenizeLogicException {
-        List<Expression<E>> tokens = new ArrayList<Expression<E>>();
-
-        int i = 0;
-        while (i < input.length()) {
-            String substring = input.substring(i);
-            char firstChar = substring.charAt(0);
-
-            if (firstChar == ' ') {
-                i += 1;
-                continue;
-            }
-            else if (firstChar == '(') {
-                tokens.add(new Paren.L<E>());
-                i += 1;
-            } else if (firstChar == ')') {
-                tokens.add(new Paren.R<E>());
-                i += 1;
-            } else if (firstChar == '!') {
-                tokens.add(new Op.Mon.Not<E>());
-                i += 1;
-            } else if (firstChar == '&') {
-                tokens.add(new Op.Bin.And<E>());
-                i += 1;
-            } else if (firstChar == '|') {
-                tokens.add(new Op.Bin.Or<E>());
-                i += 1;
-            } else {
-                // parse out the token
-                String token = this.readToken(substring);
-
-                tokens.add(factory(token));
-                i += token.length();
-            }
-        }
-
-        return tokens;
-    }
-
-    /***
      * Converts an infix logic representation into a postfix logic representation.
      * @param tokens a list of tokens in infix form.
      * @return a list of tokens in postfix (rpn) form.
@@ -378,7 +257,7 @@ abstract public class LogicExpression<E> implements Predicate<E> {
         while (scan.hasNextLine()) {
             String line = scan.nextLine();
 
-            LogicExpression<String> expr = LogicExpressions.trivial(line);
+            LogicExpression<String> expr = LogicExpressionParsers.trivial.parse(line);
 
             System.out.println("string: " + expr.toString());
             System.out.println("value:  " + expr.apply(null));
