+++ src/main/java/edu/washington/cs/knowitall/regex/RegularExpressionParsers.java
@@ -7,35 +7,33 @@ import edu.washington.cs.knowitall.regex.Expression.BaseExpression;
  *
  * @author Michael Schmitz <schmmd@cs.washington.edu>
  */
-class RegularExpressions {
+class RegularExpressionParsers {
     /**
      * Regular expressions over words where sequences are string
      * representations of words.
      */
-    public static RegularExpression<String> word(final String expr) {
-        return RegularExpression.compile(expr, new ExpressionFactory<String>() {
-            @Override public BaseExpression<String> create(final String string) {
+    public final static RegularExpressionParser<String> word =
+        new RegularExpressionParser<String>() {
+            @Override public BaseExpression<String> factory(final String string) {
                 return new BaseExpression<String>(string) {
                     @Override public boolean apply(final String token) {
                         return string.equals(token);
                     }
                 };
             }
-        });
-    }
+        };
 
     /**
      * Regular expression over characters, as in java.util.Regex.
      */
-    public static RegularExpression<Character> character(final String expr) {
-        return RegularExpression.compile(expr, new ExpressionFactory<Character>() {
-            @Override public BaseExpression<Character> create(final String string) {
+    public final static RegularExpressionParser<Character> character =
+        new RegularExpressionParser<Character>() {
+            @Override public BaseExpression<Character> factory(final String string) {
                 return new BaseExpression<Character>(string) {
                     @Override public boolean apply(final Character token) {
                         return string.equals(token.toString());
                     }
                 };
             }
-        });
-    }
+        };
 }
